+++ src/plugins/ncr/components/node-picker-field/index.jsx
@@ -12,6 +12,7 @@ import React from 'react';
 import delegate from './delegate';
 import { selectActionTypes } from '../../constants';
 import selector from './selector';
+import './styles.scss';
 
 class NodePickerField extends React.Component {
   static propTypes = {
@@ -46,7 +47,7 @@ class NodePickerField extends React.Component {
     this.state = {
       hasLoadedFirstSet: false,
       existingNodes: [],
-      menuScrollTop: 0,
+      menuListScrollTop: 0,
     };
 
     this.handleLoadMore = debounce(this.handleLoadMore.bind(this), 500, {
@@ -69,16 +70,16 @@ class NodePickerField extends React.Component {
    */
   componentDidUpdate(prevProps) {
     const { response } = this.props;
-    const { menuScrollTop } = this.state;
+    const { menuListScrollTop } = this.state;
     if (!prevProps.response || !response) {
       return;
     }
     if (
-      this.menu
+      this.menuList
       && response.get('ctx_request').get('page') > 1
       && !response.get('response_id').equals(prevProps.response.get('response_id'))
     ) {
-      this.menu.scrollTop = menuScrollTop;
+      this.menuList.scrollTop = menuListScrollTop;
     }
   }
 
@@ -119,12 +120,12 @@ class NodePickerField extends React.Component {
     if (isGetAll || !response.get('has_more')) {
       return;
     }
-    this.setState(({ existingNodes }, props) => ({
+    this.setState(({ existingNodes, menuListScrollTop }, props) => ({
       existingNodes: [
         ...existingNodes,
         ...props.response.get('nodes', []),
       ],
-      menuScrollTop: this.menu.scrollTop,
+      menuListScrollTop: get(this, 'menuList.scrollTop', menuListScrollTop),
     }), () => {
       handleSearch(response.get('ctx_request').get('q'), response.get('ctx_request').get('page') + 1);
     });
@@ -191,7 +192,6 @@ class NodePickerField extends React.Component {
                 ...props.innerProps,
                 onScroll: ({ target }) => this.handleScroll(target.scrollHeight - target.scrollTop === target.clientHeight),
               }}
-              innerRef={(e) => { this.menu = e; }}
             >
               {props.children}
             </MenuComponent>
@@ -199,6 +199,7 @@ class NodePickerField extends React.Component {
           MenuList: (props) => (
             <components.MenuList
               {...props}
+              innerRef={(e) => { this.menuList = e; }}
             >
               {props.children}
             </components.MenuList>
@@ -226,10 +227,6 @@ class NodePickerField extends React.Component {
           value: NodeRef.fromNode(node),
           node,
         }))}
-        styles={{ menuList: (base) => ({
-          ...base,
-          overflowY: 'unset',
-        }) }}
         value={value}
       />
     );