+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## patch
 * update all node pickers to request 25 per page
+* fix `nodePickerField` double scrollbar and overflow issues
 
 
 ## v0.2.3